% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_dbcon.R
\docType{methods}
\name{dbcon}
\alias{dbcon}
\alias{closeCon}
\alias{closeCon,MariaDBConnection-method}
\title{Connect to a database}
\usage{
dbcon(user, pwd = "", db = NA, host = "scidb.mpio.orn.mpg.de", path,
  driver = "MariaDB", ...)

closeCon(con)

\S4method{closeCon}{MariaDBConnection}(con)
}
\arguments{
\item{user}{username}

\item{pwd}{password ("" or NA means no password e.g. for localhost)}

\item{db}{database to connect to}

\item{host}{default to "scidb.mpio.orn.mpg.de"}

\item{path}{to credentials file (if different from the default)}

\item{driver}{MariaDB or mysql_gdal, ... Defaults to MariaDB}

\item{...}{pass to dbConnect}

\item{con}{a connection made with dbcon or \code{\link{dbConnect} }}
}
\value{
a connection object
}
\description{
\code{dbcon} returns a database connection.
                  If user and password are not given dbq looks for previously saved credentials
                  see(  \code{\link{saveCredentials}} ). \code{\link{closeCon}} closes the connection.
}
\seealso{
\code{\link{saveCredentials}}, \code{\link{dbq} }
}
