% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_dbWrite.R
\name{dbInsertInto}
\alias{dbInsertInto}
\title{dbWriteTable safe variant}
\usage{
dbInsertInto(con, name, x)
}
\arguments{
\item{con}{a *Connection object}

\item{name}{table name}

\item{x}{data.table}
}
\description{
dbWriteTable(tempTable) --> insert into select .. from --> drop tempTable
}
\examples{
x = data.table(f1 = rep('a', 10), f2 = rnorm(10), f3 = 1)
con = dbcon('mihai', host = '127.0.0.1', db = 'tests')
dbq(con, 'CREATE TABLE temp (f1 VARCHAR(50) NULL,f2 FLOAT NULL)' )
dbInsertInto(con, 'temp', x)
dbq(con, 'DROP TABLE temp')
}
