% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_dbq.R
\name{dbq}
\alias{dbq}
\title{query the database}
\usage{
dbq(con, q, ...)
}
\arguments{
\item{con}{a connection object returned by \code{\link{dbcon}}}

\item{q}{a query string. credentials are storred on disk.}
}
\value{
a data.frame or a  Spatial*DataFrame (OGR_PostgreSQLDriver) for a SELECT query, or NULL for non-SELECT queries.
}
\description{
query the database using an user-defined connection or a temp connection based on saved credentials.
}
\details{
run an SQL query.
}
\examples{
# A connection is made and used by dbq
 con = dbcon('mihai')
 d1 = dbq(con, 'SELECT * from BTatWESTERHOLZ.ADULTS')
 d2 = dbq(con, 'SELECT * from BTatWESTERHOLZ.ADULTS', enhance = TRUE)

# A temp. connection is made and closed once the date are retrieved
dbq(q = 'select now()', user = 'mihai') \%>\% str
dbq(q = 'select now()', user = 'mihai', enhance = TRUE) \%>\% str

# no return
dbq(user = 'mihai', host = 'localhost', q = 'drop database if exists this_does_not_Exists_ever_123')
dbq(con, 'drop database if exists this_does_not_Exists_ever_123')
dbDisconnect(con)
}
\seealso{
\code{\link{saveCredentials}},\code{\link{dbcon}}
}

