% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_dbWrite.R
\name{dbSafeWriteTable}
\alias{dbSafeWriteTable}
\title{dbWriteTable() on small chunks}
\usage{
dbSafeWriteTable(con, name, x, append = TRUE, chunkSize = 1000,
  verbose = TRUE, ...)
}
\arguments{
\item{con}{a *Connection object}

\item{name}{table name}

\item{x}{data.table}

\item{append}{default to TRUE}

\item{chunkSize}{default 1000 rows}

\item{...}{passed to dbWriteTable}
}
\description{
calls dbWriteTable() on small chunks
}
\examples{
x = data.table(col1 = rep('a', 1000010), col2 = rnorm(1000010))
con = dbcon('mihai', host = '127.0.0.1', db = 'tests')
dbq(con, 'CREATE TABLE temp (col1 VARCHAR(50) NULL,col2 FLOAT NULL)' )
dbSafeWriteTable(con, 'temp', x)
dbq(con, 'DROP TABLE temp')
}
