% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_dbq.R
\name{enhanceOutput}
\alias{enhanceOutput}
\title{Enhance a data.table by reference}
\usage{
enhanceOutput(d, ...)
}
\arguments{
\item{...}{optional parameters to anytime}
}
\description{
Enhance a data.table by reference
}
