% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_mysqlCLI.R
\name{my_remote2local}
\alias{my_remote2local}
\title{dumps tables between databases (wrapper to mysqldump and mysql CLI)}
\usage{
my_remote2local(db, tables, remoteUser,
  remoteHost = "scidb.mpio.orn.mpg.de", localHost = "127.0.0.1",
  localUser = "root", map = 1, call_only = FALSE, no_data = FALSE)
}
\arguments{
\item{db}{database name}

\item{tables}{tables are given as a "tableName1 tableName2".}

\item{remoteUser}{remoteUser}

\item{remoteHost}{remoteHost}

\item{localHost}{localHost Default to '127.0.0.1'}

\item{localUser}{localUser. Default to 'root'}

\item{map}{max_allowed_packet in GB (see https://dev.mysql.com/doc/refman/5.5/en/mysql-command-options.html)}

\item{call_only}{Default to FALSE}

\item{no_data}{Default to FALSE}
}
\value{
NULL or the call itself if call_only = TRUE
}
\description{
my_remote2local uses mysql CLI and mysqldump to copy tables from remote to localhost
}
\note{
non-table objects (views, procedures, etc) are exported by default.
}
\examples{
\dontrun{
my_remote2local("dbnam", "table", "user")
sysCall = my_remote2local("dbnam", "table", "user", call_only = TRUE)
system(paste ("ssh user_at_host", sysCall) )
}
}
