% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_mysqlCLI.R
\name{mysqldump_host}
\alias{mysqldump_host}
\title{mysqldump_host}
\usage{
mysqldump_host(host = "127.0.0.1", user, pwd, dir, exclude = c("mysql",
  "information_schema", "performance_schema"), parallel = TRUE)
}
\arguments{
\item{host}{default to '127.0.0.1'}

\item{user}{user}

\item{pwd}{pwd}

\item{dir}{saving location on disk.}

\item{exclude}{db-s to exclude default to c('mysql', 'information_schema', 'performance_schema')}

\item{parallel}{default to TRUE}
}
\description{
mysqldump_host
}
\examples{
\dontrun{
sdb::mysqldump_host('127.0.0.1',  user = 'mihai', dir = '~/Desktop' )

}
}
