% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_mysqlCLI.R
\name{mysqlrestore}
\alias{mysqlrestore}
\title{mysqlrestore}
\usage{
mysqlrestore(db, user = "root", filepath, call = FALSE)
}
\arguments{
\item{user}{user, default to 'root'}
}
\description{
restore sql file locally
}
\examples{
\dontrun{
db = 'BTatWESTERHOLZ'
fp = mysqldump(db, 'ADULTS BREEDING', 'mihai' , dir = tempdir() )
mysqlrestore(filepath = fp, db = db)

fp = mysqldump(db, user = 'mihai' , dir = tempdir() )
mysqlrestore(filepath = fp, db = db)
}


}

