% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_mysqlCLI.R
\name{mysqlrestore_host}
\alias{mysqlrestore_host}
\title{mysqlrestore_host}
\usage{
mysqlrestore_host(dir, host = "127.0.0.1", wipe = FALSE,
  mysqldb = FALSE, ...)
}
\arguments{
\item{dir}{a directory containing all the sql files.}

\item{host}{default to localhost}

\item{wipe}{drop all non-system db-s before restore. default to FALSE}

\item{mysqldb}{restore mysql db?. default to FALSE}

\item{...}{further options passed to mysqlrestore}
}
\description{
restore an entire db system or several db-s
}
\examples{
\dontrun{
 mysqldump_host('127.0.0.1',  'mihai', dir = '/home/mihai/Desktop' )
 mysqlrestore_host(dir = '~/Desktop/backup_127.0.0.1_09-11-19-12H' ,user = 'mihai', wipe = TRUE)
}


}
