% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_mysqlCLI.R
\name{mysqlrestoreSITE}
\alias{mysqlrestoreSITE}
\title{mysqlrestoreSITE}
\usage{
mysqlrestoreSITE(dir, filetype = ".sql.gz",
  progress = "/tmp/monitor_progress.txt", ...)
}
\arguments{
\item{dir}{a directory containing all the sql files (see mysqlrestore).}

\item{filetype}{default to .sql.gz}

\item{progress}{progress file, default to '/tmp/monitor_progress.txt'. use tail -f /tmp/monitor_progress.txt}

\item{...}{further options passed to mysqlrestore}
}
\description{
restore an entire db system or several db-s
}
\examples{
\dontrun{
mysqlrestoreSITE('/media/mihai/DATA1/scidb_backup_5.01.17/AVESatBARROW/',user = 'mihai', host = 'localhost', verbose = TRUE)
}


}

