% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeDuplicates}
\alias{removeDuplicates}
\title{removeDuplicates}
\usage{
removeDuplicates(con, table, key = "pk")
}
\description{
removeDuplicates
}
\examples{
\dontrun{
con = dbcon('mihai', host = '127.0.0.1', db = 'BTatWESTERHOLZ')
removeDuplicates(con, 'ADULTS', 'ad_pk')
}
}
