% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_sqlSnippets.R
\name{snipSave}
\alias{headQuery}
\alias{snipDrop}
\alias{snipFetch}
\alias{snipSave}
\alias{snipSearch}
\title{SQL Snippets}
\usage{
snipSave(query, description = "", user, host = "localhost")

snipFetch(id, user, host = "localhost")

snipDrop(id, user, host = "localhost")

snipSearch(kw, user, host = "localhost")

headQuery(z, nchar = 50)
}
\arguments{
\item{query}{An sql query}

\item{description}{description}

\item{user}{database user}

\item{host}{default to "localhost"}

\item{id}{snip ID}

\item{kw}{keyword}

\item{f}{A file}
}
\value{
snipSave returns the id of the saved snippet.
}
\description{
SQL Snippets
}

