% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_db}
\alias{test_db}
\title{Test database}
\usage{
test_db(user = "testuser", host = "localhost", db = "tests", pwd,
  destroy = FALSE)
}
\description{
Installs or destroys a mariadb test database
}
\note{
the functions works independently or other functions in sdb
}

