% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dbcon.R
\name{dbcon}
\alias{dbcon}
\title{connect to a database}
\usage{
dbcon(user, pwd, db = NA, host = "scidb.mpio.orn.mpg.de", path)
}
\arguments{
\item{user}{username}

\item{path}{to credentials file (if different from the default)}

\item{password}{password}

\item{database}{database to connect to}
}
\value{
a connection object
}
\description{
connects to a database
}
\details{


this functions returns a database connection. If user & password remain unspecified, it looks for credentials saved by running \code{\link{saveCredentials}}. Under Linux, the RMySQL package is used to establish the connection; under Windows, RODBC is used.
}
\seealso{
\code{\link{saveCredentials}}, \code{\link{dbq} }
}

