% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_dbcon.R
\name{dbcon}
\alias{dbcon}
\alias{closeCon}
\title{Connect to a database}
\usage{
dbcon(user, pwd = "", db = NA, host = "scidb.mpio.orn.mpg.de", path,
  driver = "MySQL", ...)

closeCon(con)
}
\arguments{
\item{user}{username}

\item{path}{to credentials file (if different from the default)}

\item{driver}{MySQL or spatial_MySQL, ... Defaults to MySQL}

\item{...}{pass to dbConnect}

\item{password}{password ("" or NA means no password e.g. for localhost)}

\item{database}{database to connect to}
}
\value{
a connection object
}
\description{
\code{dbcon} returns a database connection.
                  If user and password are not given dbq looks for previously saved credentials
                  see(  \code{\link{saveCredentials}} ). \code{\link{closeCon}} closes the connection.
}
\seealso{
\code{\link{saveCredentials}}, \code{\link{dbq} }
}
