% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dbq.R
\name{dbq}
\alias{dbq}
\title{query the database}
\usage{
dbq(con, q, ...)
}
\arguments{
\item{con}{a connection object returned by \code{\link{dbcon}}}

\item{q}{a query string. credentials are storred on disk.}
}
\value{
a data.frame for a SELECT query or otherwise NULL.
}
\description{
query the database using an user-defined connection or a temp connection based on saved credentials.
}
\details{
run an SQL query.
}
\seealso{
\code{\link{saveCredentials}},\code{\link{dbcon}}
}

