% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_dbq.R
\name{dbq}
\alias{dbq}
\title{query the database}
\usage{
dbq(con, q, geom, ...)
}
\arguments{
\item{con}{a connection object returned by \code{\link{dbcon}}}

\item{q}{a query string. credentials are stored on disk.}
}
\value{
a data.frame or a  Spatial*DataFrame (spatial_MySQL) for a SELECT query, or NULL for non-SELECT queries.
}
\description{
query the database using an user-defined connection or a temp connection based on saved credentials.
}
\details{
run an SQL query.
}
\examples{
\dontrun{
# A connection is made and used by dbq
 con = dbcon('testuser', host =  '127.0.0.1', pwd =  askpass::askpass() )
 d1 = dbq(con, 'SELECT * from tests.t1')
 d2 = dbq(con, 'SELECT * from tests.t1', enhance = TRUE)

# A temp. connection is made and closed once the data is retrieved
dbq(q = 'select now()', user = 'testuser', host =  '127.0.0.1', pwd =  'cs') 
dbq(q = 'select now()', user = 'testuser',host =  '127.0.0.1', pwd =  'cs',  
	enhance = TRUE)\%>\% print

# spatial return
s = dbq(con, q = 'select * from tests.t3', geom = 'SHAPE')
s = dbq(con, q = 'select SHAPE from tests.t3', geom = 'SHAPE')
s = dbq(con, q = 'select * from tests.t2', geom = 'SHAPE')

}
}
\seealso{
\code{\link{saveCredentials}},\code{\link{dbcon}}
}
