% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_dbq.R
\name{dbq}
\alias{dbq}
\title{query the database}
\usage{
dbq(con, q, ...)
}
\arguments{
\item{con}{a connection object returned by \code{\link{dbcon}}}

\item{q}{a query string. credentials are stored on disk.}
}
\value{
a data.frame or a  Spatial*DataFrame (spatial_MySQL) for a SELECT query, or NULL for non-SELECT queries.
}
\description{
query the database using an user-defined connection or a temp connection based on saved credentials.
}
\details{
run an SQL query.
}
\examples{
# A connection is made and used by dbq
 con = dbcon('mihai', host =  '127.0.0.1')
 d1 = dbq(con, 'SELECT * from BTatWESTERHOLZ.ADULTS')
 d2 = dbq(con, 'SELECT * from BTatWESTERHOLZ.ADULTS', enhance = TRUE)

# A temp. connection is made and closed once the data is retrieved
dbq(q = 'select now()', user = 'mihai', host =  '127.0.0.1') \%>\% str
dbq(q = 'select now()', user = 'mihai',host =  '127.0.0.1',  enhance = TRUE) \%>\% str

# null return
dbq(user = 'mihai', host =  '127.0.0.1', q = 'set @c=1')
dbq(con, 'set @c=1')
dbDisconnect(con)

# spatial return
con = dbcon('mihai', host =  '127.0.0.1', db = 'tests', driver = 'spatial_MySQL')
s = dbq(con, q = 't3')
s = dbq(con, q = 'select * from t3 limit 1')

con = dbcon('mihai', host =  '127.0.0.1', db = 'AVES_ranges', driver = 'spatial_MySQL')
s = dbq(con, q = "select * from breeding_ranges_v1 where scinam = 'Parus major'")
}
\seealso{
\code{\link{saveCredentials}},\code{\link{dbcon}}
}

