% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mysql_remote2local.R
\name{my_remote2local}
\alias{my_remote2local}
\title{dumps tables on localhost}
\usage{
my_remote2local(db, tables, pwdLocal, userLocal = "root")
}
\arguments{
\item{tables}{tables are given as a "tableName1 tableName2".}
}
\description{
my_remote2local uses mysql CLI and mysqldump to copy tables from remote to localhost
}
\note{
non-table objects (views, procedures, etc) are exported by default.
}
\seealso{
\code{\link{getCredentials}},\code{\link{mysqlCLI}}
}

