% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_mysqlCLI.R
\name{mysqldump}
\alias{mysqldump}
\title{mysqldump}
\usage{
mysqldump(db, tables, user, host = "scidb.mpio.orn.mpg.de", filenam,
  dir = getwd(), call = FALSE)
}
\arguments{
\item{db}{db}

\item{tables}{tables are given as a "tableName1 tableName2".}

\item{user}{user}

\item{host}{default to 'scidb.mpio.orn.mpg.de'}

\item{call}{show mysql call}
}
\value{
the file path to the sql file
}
\description{
mysqldump
}
\examples{
\dontrun{
fp = mysqldump('BTatWESTERHOLZ',  user = 'mihai', dir = tempdir() )
mysqldump('BTatWESTERHOLZ', 'ADULTS BREEDING', 'mihai' , dir = tempdir() )
}


}
