% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_mysqlCLI.R
\name{mysqldump}
\alias{mysqldump}
\title{mysqldump}
\usage{
mysqldump(db, tables, user, pwd, host = "scidb.mpio.orn.mpg.de", filenam,
  dir = getwd(), dryrun = FALSE, ...)
}
\arguments{
\item{db}{db}

\item{tables}{tables are given as a "tableName1 tableName2".}

\item{user}{user}

\item{pwd}{pwd}

\item{host}{default to 'scidb.mpio.orn.mpg.de'}

\item{filenam}{filenam. If missing then constructed from Sys.time()}

\item{dir}{saving location on disk.}

\item{dryrun}{when TRUE return call only}

\item{...}{further arguments to mysqldump (e.g. --no-data --no-create-db)}
}
\value{
the file path to the sql file
}
\description{
mysqldump
}
\examples{
\dontrun{
fp = mysqldump('BTatWESTERHOLZ',  user = 'mihai', dir = tempdir() )
mysqldump('BTatWESTERHOLZ', 'ADULTS BREEDING', 'mihai' , dir = tempdir() )
}


}
