% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_mysqlCLI.R
\name{mysqlrestore}
\alias{mysqlrestore}
\title{mysqlrestore}
\usage{
mysqlrestore(file, db, user = "root", host = "127.0.0.1", dryrun = FALSE)
}
\arguments{
\item{file}{sql or sql.gz file}

\item{db}{database name}

\item{user}{user, default to 'root'}

\item{host}{default to  '127.0.0.1'}

\item{dryrun}{only print the mysql cli call and exit}
}
\description{
restore sql file locally
}
\examples{
\dontrun{
db = 'BTatWESTERHOLZ'
fp = mysqldump(db, 'ADULTS BREEDING', 'mihai' , dir = tempdir() )
mysqlrestore(file = fp, db = db)

fp = mysqldump(db, user = 'mihai' , dir = tempdir() )
mysqlrestore(file = fp, db = db)
}


}
