% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_mysqlCLI.R
\name{mysqlrestore}
\alias{mysqlrestore}
\title{mysqlrestore}
\usage{
mysqlrestore(file, db, user, host = "127.0.0.1", dryrun = FALSE)
}
\arguments{
\item{file}{sql or sql.gz file}

\item{db}{database name}

\item{user}{user, default to 'root'}

\item{host}{default to  '127.0.0.1'}

\item{dryrun}{only print the mysql cli call and exit}
}
\description{
restore sql file locally
}
\examples{
\dontrun{
fp =  mysqldump(db = 'tests', table='tests', user='testuser' , dir = tempdir(), compress = FALSE )
mysqlrestore(file = fp, db = 'tests', user = 'testuser')

fp =  mysqldump(db = 'tests', table='tests', user='testuser' , dir = tempdir(), compress = TRUE )
mysqlrestore(file = fp, db = 'tests', user = 'testuser')

}


}
