% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_mysqlCLI.R
\name{mysqlrestoreSITE}
\alias{mysqlrestoreSITE}
\title{mysqlrestoreSITE}
\usage{
mysqlrestoreSITE(dir, filetype = ".sql.gz", exclude = c("mysql",
  "information_schema", "performance_schema"),
  progress = "/tmp/monitor_progress.txt", ...)
}
\arguments{
\item{dir}{a directory containing all the sql files (see mysqlrestore).}

\item{filetype}{default to .sql.gz}

\item{exclude}{db-s to exclude default to c('mysql', 'information_schema', 'performance_schema')}

\item{progress}{progress file, default to '/tmp/monitor_progress.txt'. use tail -f /tmp/monitor_progress.txt}

\item{...}{further options passed to mysqlrestore}
}
\description{
restore an entire db system or several db-s
}
\examples{
\dontrun{
mysqlrestoreSITE('..../scidb_backup_5.01.17/',user = 'mihai', host =  '127.0.0.1', verbose = TRUE)
}


}
