% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{probeDB}
\alias{probeDB}
\title{Probe a db server}
\usage{
probeDB(probe = TRUE, name = "scidb.mpio.orn.mpg.de", port = 3306)
}
\arguments{
\item{probe}{if FALSE, will not probe the db and return ''127.0.0.1'. Default to TRUE.}

\item{name}{default to scidb.mpio.orn.mpg.de}

\item{port}{default to 3306 (mariadb's default port )}
}
\value{
when reachable return the name of probed server else falls back to localhost (127.0.0.1)
}
\description{
Fast probe a db server see if it can be reached through a given port. Default to scidb.mpio.orn.mpg.de and port 3306
}
\examples{
\dontrun{
probeDB()

}
}
