% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeDuplicates}
\alias{removeDuplicates}
\title{removeDuplicates}
\usage{
removeDuplicates(con, table, key = "pk")
}
\arguments{
\item{con}{con}

\item{table}{table}

\item{key}{key. default to 'pk'.}
}
\description{
removeDuplicates
}
\examples{
\dontrun{
con = dbcon('mihai', host = '127.0.0.1', db = 'tests')
removeDuplicates(con, 'n1', 'id')
}
}
