% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{package}
\name{sdb}
\alias{sdb}
\title{\strong{s}cidb \strong{d}ata \strong{b}ase interface}
\description{
An interface to \href{http://scidb.mpio.orn.mpg.de}{scidb.mpio.orn.mpg.de} databases
}
\examples{
\dontrun{
test_db(pwd = askpass::askpass() ) # make it available for tests
}

}
\author{
Mihai Valcu \email{valcu@orn.mpg.de}
}
