% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_db}
\alias{test_db}
\title{Test database}
\usage{
test_db(user = "testuser", host = "localhost", db = "tests",
  destroy = FALSE)
}
\description{
Installs or destroys a mariadb test database
}
\note{
credentials should be stored in ~/.my.cnf which should contain
[client]
user = 'username'
password = 'pwd'
}
\examples{
/donotrun{
test_db('mihai')
}

}

