% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_db}
\alias{test_db}
\title{Test database}
\usage{
test_db(user = "testuser", host = "127.0.0.1", db = "tests", pwd,
  destroy = FALSE)
}
\description{
Installs or destroys a mariadb test database
}
\note{
test_db() works independently on other functions in sdb
}
